#  ostree repo commit: Speed up composing trees with `--tree=ref`