--- a/src/libostree/libostree-devel.sym
+++ b/src/libostree/libostree-devel.sym
@@ -22,6 +22,7 @@ LIBOSTREE_2018.7 {
 global:
     ostree_mutable_tree_fill_empty_from_dirtree;
     ostree_mutable_tree_new_from_checksum;
+    ostree_mutable_tree_check_error;
 } LIBOSTREE_2018.6;
 
 /* Stub section for the stable release *after* this development one; don't