@jlebon 2018-06-25
bot, add author to whitelist