@wmanley 2018-06-25
I've been doing some more fiddling with this and I think there is a bug in here still.  I'm trying to create a minimal testcase that reproduces the symptom.