@cgwalters 2018-06-25
See https://github.com/ostreedev/ostree/issues/1569 - I just did a bit of digging and added a commit link there.