@wmanley 2018-06-25
I've figured out what the bug is: `ostree_mutable_tree_get_contents_checksum` expects that one of it's subdirs' `contents_checksum` will be NULL if it's not valid.  This isn't the case for LAZY mtrees, because we're certain of the checksums.