@wmanley 2018-06-26
Ok, tests are passing now, including the new test I wrote for the bug I found in my previous implementation.  The diff is a little more complex now that I've had to add the parent pointer, but I think it all makes sense.