@wmanley 2018-07-02
I've rebased (and squashed) this since #1655 was merged.