@wmanley 2018-07-05
> Can you add some unit tests for this patch too? Or is that too hard to do without lowering the higher level usage too? If that's the case I'd probably prefer having the changes as a single PR.

What did you have in mind?  I added some unit tests for this functionality in 36f7a6d18267814b4d11cb5a89e2ed7993b3d543 which was already merged as part of #1645.

I suppose I could write some tests in C which explicitly call `ostree_mutable_tree_new_from_checksum` and `ostree_mutable_tree_fill_empty_from_dirtree`?