@cgwalters 2018-07-05
> What did you have in mind? I added some unit tests for this functionality in 36f7a6d which was already merged as part of #1645.

Let's back up a second here; in this scenario you mention in the commit message:

> ostree commit --tree=ref=base-rootfs --tree=ref=container1 --tree=ref=container2

Are the trees disjoint?  For example `base-rootfs` looks like a traditional system with an empty `/usr/share/containers` directory, then `container1` looks like `/usr/share/containers/container1/usr/*files here*`, `container2` looks like `/usr/share/containers/container2/usr/*files here*`, etc.

If I'm understanding the optimization here, it's really predicated on that.

Now...I'm staring at that test case code and only sort of understanding it.  It looks like it's more testing combinations of `--tree=dir` and `--tree=ref` rather than e.g. disjointness of `--tree=ref`, right?