@cgwalters 2018-07-05
(Also I pushed a fixup for the error handling, let me know what you think)