@cgwalters 2018-07-05
To rephrase some of the musings from #1569 - in your use case (as well as the promotion case), it feels like we basically want to make it a fatal error if we're not hitting the fast path right?

I completely get that we need to support mixing `--tree=ref` and `--tree=dir` because we initially exposed such as flexible API.  Hmm.  So in the above issue I proposed a `--from-ref-contents` which would be like `--tree=ref`, but you actually need a generalization of that.

How about something like this:

`ostree commit --disjoint-union --ref=base-os:/ --ref=container1:/usr/share/containers/container1 --ref=container2:/usr/share/containers/container2`

The args to `--ref` are `REF:SUBPATH`.

One (IMO) nicer thing about this is that the path prefix is specified at commit time, so one can have the `container1` tree just be rooted at `/` if that makes sense?  Makes it nicer to `ostree ls` them for example.

And further it would be a fatal error if the refs overlapped at all.